refactor: refactored DetectionGenerator into proper DataLoader
