+++ doctr/datasets/loaders/__init__.py
@@ -1,2 +1,1 @@
-from .detection import *
 from .recognition import *
